Don't attempt to use Blob URL if the file is larger than 100MB
Fixes https://github.com/feross/render-media/issues/18

If a video file doesn't work with videostream or mediasource
approaches, it doesn't make sense to fallback to the Blob URL approach
when the file is large (>100MB), because that requires the entire file
to load upfront, which will take a really long time.

Let's just set a sanity cap on the largest size.